backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
groq==0.4.1
faster-whisper==1.0.1
feedparser==6.0.10
requests==2.31.0
pydub==0.25.1
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
